refactor(api): extraer helpers compartidos de las rutas (#187) — fix #186
