+++ app/api/undo/route.ts
@@ -1,30 +1,9 @@
-import { timingSafeEqual } from 'crypto'
 import { revalidateTag } from 'next/cache'
 import { NextRequest, NextResponse } from 'next/server'
-import { get, put } from '@vercel/blob'
+import { put } from '@vercel/blob'
+import { htmlResponse, readBlobJson, secretsMatch } from '@/lib/api-helpers'
 
 export const maxDuration = 30
-
-function secretsMatch(provided: string, expected: string): boolean {
-  const a = Buffer.from(provided)
-  const b = Buffer.from(expected)
-  if (a.length !== b.length) return false
-  return timingSafeEqual(a, b)
-}
-
-async function readBlobJson(pathname: string): Promise<unknown[] | null> {
-  const result = await get(pathname, { access: 'private', useCache: false })
-  if (!result) return null
-  const text = await new Response(result.stream).text()
-  return JSON.parse(text) as unknown[]
-}
-
-function htmlResponse(title: string, body: string): NextResponse {
-  return new NextResponse(
-    `<!doctype html><html><body style="font-family:system-ui,sans-serif;text-align:center;padding:64px"><h1>${title}</h1><p>${body}</p></body></html>`,
-    { headers: { 'Content-Type': 'text/html; charset=utf-8' } },
-  )
-}
 
 export async function GET(req: NextRequest): Promise<NextResponse> {
   const secret = req.nextUrl.searchParams.get('secret') ?? ''
